cmake: Fix, modernize and simplify build-system
This commits updates the build-system so that the CLI can be successfully
bundled within a Slicer extension, a Slicer custom application or as a
standalone project.

* update minimum required version from CMake 2.8 to 3.5

* rename CLI from RigidWrapper to RigidAlignment

* rename header RigidAlignment.h to RigidAlignmentImpl.h where Impl
  stands for "Implementation"

* remove the extra static library "RigidAlignment" and directly include the
  source in the CLI.

* strip trailing white spaces

* convert tabs to spaces
