--- src/CMakeLists.txt
@@ -1,4 +1,0 @@
-add_library(RigidAlignment_SOURCES 
-        STATIC
-        RigidAlignment.cpp)
-target_link_libraries(RigidAlignment_SOURCES ${Mesh_LIBRARIES})
