--- wrapper/CMakeLists.txt
@@ -1,20 +1,0 @@
-if(UNIX)
-    set( INSTALL_RUNTIME_DESTINATION bin)
-    set( INSTALL_LIBRARY_DESTINATION lib)
-    set( INSTALL_ARCHIVE_DESTINATION lib)
-endif()
-
-SEMMacroBuildCLI(
-		NAME RigidWrapper
-		EXECUTABLE_ONLY
-		TARGET_LIBRARIES ${Mesh_LIBRARIES} RigidAlignment_SOURCES ${VTK_LIBRARIES}
-		INCLUDE_DIRECTORIES ${CMAKE_CURRENT_SOURCE_DIR} 
-		RUNTIME_OUTPUT_DIRECTORY ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}
-    	LIBRARY_OUTPUT_DIRECTORY ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}
-    	ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_ARCHIVE_OUTPUT_DIRECTORY}
-    	INSTALL_RUNTIME_DESTINATION ${INSTALL_RUNTIME_DESTINATION}
-    	INSTALL_LIBRARY_DESTINATION ${INSTALL_LIBRARY_DESTINATION}
-    	INSTALL_ARCHIVE_DESTINATION ${INSTALL_ARCHIVE_DESTINATION}
-)
-
-install(TARGETS RigidWrapper RUNTIME DESTINATION ${INSTALL_RUNTIME_DESTINATION})
