--- wrapper/RigidWrapper.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<executable>
-<title>RigidAlignment</title>
-<version>1.0</version>
-<contributor>Ilwoo Lyu</contributor>
-<description>Rigid alignment of the landmarks on the unit sphere: the input models share the same unit sphere and their landmarks are defined as spherical coordinates (x, y, z) or vertex indices</description>
-	<parameters>
-		<string>
-			<longflag>sphere</longflag>
-			<flag>s</flag>
-			<name>sphere</name>
-			<description>Template sphere model</description>
-	    </string>
-		<directory>
-			<longflag>landmark</longflag>
-			<flag>l</flag>
-			<name>landmark</name>
-			<description>Landmark directory</description>
-		</directory>
-
-		<directory>
-			<longflag>mesh</longflag>
-			<!-- <flag></flag> -->
-			<name>mesh</name>
-			<description>Mesh directory</description>
-		</directory>
-
-		<geometry fileExtensions=".csv">
-			<longflag>inputCSV</longflag>
-			<name>inputCSV</name>
-			<default>0</default>
-			<description>Input CSV file</description>
-	    </geometry>
-
-		<directory>
-			<longflag>output</longflag>
-			<flag>o</flag>
-			<name>output</name>
-			<description>Output path</description>
-	    </directory>
-	    <directory>
-			<longflag>lmindex</longflag>
-			<name>outputLM</name>
-			<description>Output path of the vertex indices of the landmarks (optional)</description>
-	    </directory>
-	    <boolean>
-	    	<longflag>lmcoord</longflag>
-	    	<name>lmtype</name>
-			<description>regards landmarks as spherical coordinates</description>
-	    </boolean>
-  	</parameters>
-</executable>
